chore: common폴더를 shared로 이름 변경 후 파일 이동
