--- src/constants.ts
@@ -1,1 +1,0 @@
-export const BASE_URL = "https://woowa.yiheon.com";
