+++ src/feature/CardRegisterPage/utils/cardCompany.ts
@@ -1,4 +1,4 @@
-import type { CardCompanyType } from "../../../common/types/CardCompany";
+import type { CardCompanyType } from "../../../shared/types/CardCompany";
 import { CARD_COMPANY } from "../../CardListPage/constants";
 import type { IssuerCodeType } from "../../CardListPage/types/CardCompay";
