+++ src/shared/constants.ts
@@ -1,0 +1,15 @@
+export const BASE_URL = "https://woowa.yiheon.com";
+
+import type { CardCompanyType } from "../shared/types/CardCompany";
+
+export const cardColors: Record<CardCompanyType | "default", string> = {
+  default: "#333333",
+  BC카드: "#F04651",
+  신한카드: "#0046FF",
+  카카오뱅크: "#FFE600",
+  현대카드: "#000000",
+  우리카드: "#007BC8",
+  롯데카드: "#ED1C24",
+  하나카드: "#009490",
+  국민카드: "#6A6056",
+};
