+++ src/shared/hooks/useInputErrorState.tsx
