+++ src/shared/types/CardCompany.ts
