+++ src/shared/types/CardInfoType.ts
